Handler-less kernel submit path (parallel_for with nd_range) (#19294)
This PR introduces a fully handler-less kernel submission path. The
feature is not complete yet. For testing purposes we introduce the
__DPCPP_ENABLE_UNFINISHED_NO_CGH_SUBMIT macros to enable unit tests for
the new handler-less path. This macro should not be used by the
application, and a legacy handler-based path is used. Once the
handler-less path is fully implemented, we will switch corresponding
APIs to use it unconditionally and will remove the macros.

This PR covers:
1. A parallel_for, nd_range based kernel submit.
2. Parallel_for queue shortcut, enqueue_functions extension and KHR free
functions extension.
3. A scheduler-based kernel submission.
4. A new unit test which covers the host task and kernel ordering for an
in-order queue (including the handler-less path).