[SYCL][SYCLBIN] Select best images based on state (#20197)
Previously, the implementation would prioritize native device code
binaries over IR modules indiscriminately. However, for input state and
object state, native device code images are unsuitable. As such, this
commit makes it so that the native device code images are only selected
when the target state is executable.

As an effect of this change, input and object tests will no longer run
for CUDA and HIP. This is appropriate given these targets only currently
support generating native device code images, which are meant to be as
close to executable state as possible. Note that the driver doesn't
currently make such limitations of the generated SYCLBIN files, so
producing input and object state binaries with CUDA and HIP targets will
result in kernel bundles with no binary images.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>