[UR][L0][L0v2] Explicitly mark timestamped events
This is a workaround for a case where L0 driver fails to retrieve a
timestamp but doesn't return an error. In such case, when event
profiling info is queried we check the timestamp value and assume the
event was not timestamped even though it was.
This scenario can happen on newer platform with not yet full driver
support.