[SYCL] Don't return last event in `ext_oneapi_submit_barrier`  (#20235)
This is a cherry-pick of intel/llvm#20159 (and intel/llvm#20241)

**Problem**

As an optimization in `ext_oneapi_submit_barrier`, for in-order queue,
when waitlist is empty, we return the last event, if it is available.
However, after https://github.com/intel/llvm/pull/18277, we no longer
store last event and instead submit a marker to get the last event. And
this caused performance regression in some of our internal benchmarks.

**Solution**
Don't return last event and instead submit a barrier with empty
waitlist.


Patch-by: Udit Kumar Agarwal <udit.agarwal@intel.com>