[Benchmark] Disable sycl-bench/lin_reg_coeff benchmark (#20156)
`lin_reg_coeff` seems to be sporadically causing a GPU hang on the BMG
runner: The hang is so bad that we are forced to restart the machine
when the hang happens.

Undoubtedly there is probably a driver issue involved here
(compute-runtime 25.31.34666.3 is used), but to ensure the CI runs
continuously I am disabling the test.