[SYCL][Graph] Remove UR sync point tracking in linear graphs (#20291)
When the command buffer in a partition is created with property
`isInOrder`, explicit sync point dependencies are ignored by UR. We
currently set this property when the graph is linear but still create
and store sync points. We should avoid tracking sync points in this case
as it adds unnecessary overhead to graph finalization.