[CI] Fail benchmark CI only on regressions in CPU instruction count (#20277)
Benchmark script changes in this PR are NFC.

Bit of a long overdue change: Wall time has proven to not be a good
metric for stable performance numbers. This PR changes SYCL benchmarking
CI regression filters to only fail on severe regressions in tests
measuring CPU instruction count only.

This PR does not affect the behavior of benchmarking scripts. It only
affects the behavior of the CI w.r.t. regression failures for _SYCL_: it
does not concern regressions detected in UR or L0.