[NFC][ClangLinkerWrapper] fix tests after 65d730b4a5a (#32475)
These new tests are added after 65d730b4a5a
However, the logic in xmain are still very different from upstream,
eg: we skip bundleLinkedOutput for OFK_SYCL,
so the dry_run logic for test is not even called.

We also have more logic so we require sycl device lib to be not empty,
add the fake lib to get it pass.