[NFC][HIP] XFAIL amdgpu-builtin-in-lambda.hip (#32529)
Looks like it might be a bug in AMDGPU Sema.

Due to:

commit:
[efa907e899339f1c2d9b885202ec23ed612300ed](https://github.com/intel-restricted/applications.compilers.llvm-project/commit//efa907e899339f1c2d9b885202ec23ed612300ed)
for amdgpu-builtin-in-lambda.hip

Author: Juan Manuel Martinez Caamaño
[Clang][AMDGPU] Defer amdgcn.*to_lds builtins size diagnostics (#160140)

Before, the diagnostic was emitted immediately, as soon as the error was
detected. This is problematic during the host compilation, since
the compiler performs semantic analysis of `_device_` functions with
the host's
target attributes.

A solution for this is to use `SemaRef.targetDiag` to defer the
diagnostic. The diagnostic will then be printed only if the function is
emitted.

The test included in this patch highlights a second problem: we cannot
compile a file having a call to `__builtin_amdgcn_load_to_lds` on a
`_device_` function since we typecheck the signature.
The issue is that, `_shared_ void*` on X86 doesn't translate to
`addrspace(3) void*`, so the compilation fails.

Fixes: SWDEV-555501