[SYCL] Use %t.dir for DeviceImageDependencies tests (#20372)
Last pulldown pulled the deprecation of `%T` for llvm-lit, and
DeviceImageDependencies tests were using it. It was reported that
`%t.dir` wasn't working for these tests in the pulldown PR
(https://github.com/intel/llvm/pull/20343#discussion_r2426535174), and a
workaround was made to make them pass provisionally. This patch moves to
`%t.dir` as it's the standard way after `%T` removal.