Always emit coopmat conversions as SPIR-V friendly IR calls (#3398)
Map all cooperative matrix type conversions to SPIR-V friendly IR calls,
regardless of the environment specified.

In particular, do not attempt to map such conversions to the OpenCL
`convert` builtin. The SPIR-V TargetExtType is already encoded in the
function suffix, so the previous translation was an odd hybrid between
OpenCL and SPIR-V friendly IR.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/9d56f01d5589998