[SYCL] Implement `sycl_khr_includes` extension (#20339)
The specification can be found in KhronosGroup/SYCL-Docs#814

The implementation is a non-functional change for the existing codebase,
but definitions of macro documented by the SYCL 2020 specifications were
moved around a little bit to provide them all through the new
`sycl/khr/version.hpp`.

New headers are always present and their content is always available.
However, we do not define `SYCL_KHR_INCLUDES` macro yet because the
extension is still on review. That indicates that it is not formally
supported by the implementation.