[SYCL] Implement coverage instrumentation for device code (#20206)
This PR extends Clang's source-based code coverage to work with SYCL
device code. It includes the following changes:

1. The `InstrProfilingLoweringPass` was updated to lower profiling
counters to SYCL device globals.
2. A new function was added to the compiler runtime to increment the
host-side profiling counters.
3. The SYCL runtime was updated to send the contents of the device-side
profiling counters to the new compiler runtime function when their
device global map entries are being freed.

This feature may not work correctly for functions that differ between
host and device due to the use of the `__SYCL_DEVICE_ONLY__` macro. In
such cases, it may not be possible to correlate the profiling counters
from the device to the host. Resolves #7803.

---------

Signed-off-by: Michael Aziz <michael.aziz@intel.com>
Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>
Signed-off-by: Michael Aziz <michael.aziz@intel.com>