[Benchmarks] remove force rebuild path, fail on error in CI, shorten CI time (#20388)
## Change in CI

### Add exit-on-failure to benchmarks CI run
There was exit-on-failure option available in benchmark scripts, but it
was not used by CI that checks correctness. Now it is so we catch more
errors in CI.

### Speed up benchmarks CI
CI was running as many interations as in normal performance measurement,
which was pointless for just checking scripts correctness. Now
iterations count is limited to minimum which changed benchmarks CI time
from 1 hour to 15 minutes.

## Change in Benchmark scripts

### disabled one failing benchmark, fixed another

Once CI exits on any failure it revealed two errors in benchmarking. One
was easy to fix. Another will be done in the future and now its
benchmarks are disabled

### removed needs_rebuild argument

Defining whether project use install dir is in one place - GitProject
constructor, not in configure and needs_rebuild arguments which was
error prone. No force rebuild path any more.

"--no-rebuild" option to devops/scripts/benchmarks/main.py is removed
and the behavior of the script is the same as previously was with that
option. This simplifies the code

Previously --no-rebuild option caused compute-benchmarks, llama , etc...
dependent project to be not rebuild if already compiled in proper
version (commit hash hardcoded in benchmarks scripts matches).

Now this behavior is default and the only one.

if suite is enabled by options (--sycl, --ur, etc..), compiled and its
tag (e.g. compute-benchmarks tag) matches benchmarks's hardcoded one -
it is not rebuild, otherwise it is rebuild.

In CI simply workdir is cleaned before running main.py, so previous
behavior is preserved.

During manual work, if one wants to force a rebuild a dependent repo
he/she needs to remove its binary dir from workdir.