[SYCL][NFC] Drop confusing `std::move` (#20455)
Input argument for this `std::move` is `const &`, meaning that the
result is `const &&`. `std::vector` does not have a move constructor
from `const &&` and therefore it fallbacks to regular copy constructor
taking `const &`.

Dropped that `std::move` to better describe what actually happens there.
As a side effect, this change should help Coverity to stop stubmling
upon this and complaining that some object is used after it was moved
(see CID `535422`).