[SYCL][Doc] Revamp proposed spec for USM shortcuts (#20468)
After further discussion, it seems that our customer wants USM shortcut
functions that take a `device` parameter, but infer the context from
that device.  Of course, this works only for a subset of the USM
allocations functions (and not for the deallocation or query functions
at all).  Add a new proposed specification with these shortcuts.

The previous shortcuts (those that infer both the device and the
context) are moved to a different proposed extension in case we want to
revive that in the future.

This commit also fixes some typos in the previous proposed spec.
I also discovered that shortcuts for the "parameterized allocation
functions" were inadvertently omitted from the previous proposed spec,
so add those too.