[E2E] Enable end-to-end tests supported by the SPIR-V backend  (#20418)
This PR re-enables SYCL end-to-end tests previously marked as XFAIL or
UNSUPPORTED on the SPIR-V backend, as they now pass. For tests that are
still failing, new issues have been created in llvm-project and added as
XFAIL-TRACKER.

This is the second PR for re-enabling SYCL end-to-end because the
previous one is reverted due to post-commit failure.