[SYCLLowerIR] Fix a bug in hierarchical parallelism implementation (#20484)
When inside a `parallel_for_work_group` context, calls to functions that
contain calls to `parallel_for_work_item` are being lowered incorrectly
into IR in that they are being put under the work-group leader branch in
the IR which is semantically incorrect as this function should be called
in every work item. This manifests when we have an indirect function
call to `parallel_for_work_item` together with at least one other direct
call to `parallel_for_work_item` in the same `parallel_for_work_group`
context and it leads to a program that hangs. This PR fixes the issue
and adds a couple of other tests to check this behavior.