[SYCL][NFC] Avoid temporary object creation when do tracing (#20443)
Change tracing functions to avoid temporary `std::string_view` object
creation. The issue was that if a string literal is passed as a `Msg`
parameter, the temporary `std::string_view` object was created. When the
constructor of the `std::string_view` is called, internally it
calculates the length of the C-string. This PR avoids that.