[Clang][SYCL] Fix dangling pointers and ODR violation in free functions (#20422)
**Problems**
Problem 1: When a library consisting of free function kernels is
registered with SYCL RT, we store pointers (as `string_view`) to free
function names in `m_FreeFunctionKernelGlobalInfo` but we do not remove
them from `m_FreeFunctionKernelGlobalInfo` when the library is unloaded.
Thus, we end up holding dangling pointers and any further operation on
`m_FreeFunctionKernelGlobalInfo` might segfault.

Problem 2: Consider the case when you have multiple TUs with free
functions and they are compiled separately but linked together into a
single shared lib. In that case, we will have multiple definition of
`static GlobalMapUpdater updater` in the shared lib => violating ODR

**Solution**
Discard pointers to free function names when library is unloaded and
have `GlobalMapUpdater` defined in anonymous namespace, instead of
`sycl::v1::detail`

---------

Co-authored-by: premanandrao <premanand.m.rao@intel.com>