[clang][SYCL] Fix field/base builtins for PCH support
We have added several builtins for device copyability checking a while
back. These are implemented via new AST nodes, so they require
implementation of ASTReaderStmt's and ASTWriterStmt's methods. The
implementation was in place but it was untested so we never saw that it
is missing reading of the builtin expression's type which resulted in
crashes during PCH reading.