[CI] Don't rely on `github.event_name` in `devops/actions/run-tests/e2e/action.yml`
This effectively reverts

"[CI] Use E2E container binaries in `run-only` manual E2E task dispatch (#19744)"

The problem is that `github.event_name` corresponds to the top-level workflow,
not how `.github/workflows/sycl-linux-run-tests.yml` was triggered, effectively
doing the wrong thing for the manual dispatch of (at least) post-commit workflow.

Instead of that, add a new fake input parameter for the `workflow_call` trigger
in `sycl-linux-run-tests.yml` to use as the condition.