[SYCL][E2E] Workaround enum parameter free function kernel test failure (#20423)
We recently disabled dead arguments elimination optimization
for free function kernels and most likely this triggered this issue.

Default-constructed accessors used in the test only matter during FE
stage when we instantiate necessary templates, but at LLVM IR level
they are completely unused, because we do not access their fields or
call any member functions.

As such, with dead arguments elimination optimization disabled, SYCL
RT stopped to ignore them and thus we uncovered a bug there with
use-after-free.

Considering that the test was aimed to test a different thing, a workaround
is made to avoid sporadic failures of the test. #20225 was submitted to
track the underlying issue.

Most likely some kind of reference to the accessor is preserved within
`handler` whose lifetime may be longer than the one of the accessor,
thus causing `finalize` method to access a dead memory region.