[SYCL][clang-linker-wrapper] Update --bitcode-library option for SYCL (#20331)
This PR updates the `--bitcode-library` option in `clang-linker-wrapper`
to link bitcode device libraries that are used in NVIDIA and AMD GPU
targets in SYCL offloading. This option had some previous behavior, but
it was unused, so I deleted it and repurposed it for SYCL.

`SYCL::getDeviceLibraries()` returns both fat objects and bitcode device
libraries mixed together, while the `-sycl-device-libraries` option only
handles fat objects. This was causing `clang-linker-wrapper` to not link
the `libdevice` for nvptx/AMD targets.

This PR also removes the `-sycl-nvptx-device-libraries` option, which
seemed to be doing the same thing but only for nvptx targets.