[SYCL][NewOffloadModel] Enforce default SYCL device library to use old offloading model (#20569)
This pull request enforces the default SYCL device library to always be
built with the old offloading model, regardless of which offloading
model (new or old) is set as the default. This is implemented by adding
`--no-offload-new-driver` to the `device_compile_opts` in
`libdevice/cmake/modules/SYCLLibdevice.cmake`.
The motivation for this change is to address test failures in
`Config/kernel_from_file.cpp` and `SeparateCompile/test.cpp` that occur
when the new offloading model is enabled as the default. These failures
happen because `clang-offload-bundler` does not support extracting
libraries built with the new offloading model.