[NewOffloadModel] Fix transition of compile options from image to SYCL Offload Wrapper (#19579)
This patch fixes the handling of compilation and link options encoded in
images. Also, it fixes option redirection to the backend compiler and to
the offload wrapper depending on the Triple (AOT or JIT). In this patch,
clang-linker-wrapper extracts and processes compilation and link options
separately for every target.

Compilation and linking options are moved from `SYCLWrappingOptions` to
`SYCLImage` because different targets may have different sets of
options.

The following command line options are removed from clang-linker-wrapper
because compilation and link options are per-target which requires to
encoding them in images.
* sycl_backend_compile_options_EQ 
* sycl_target_link_options_EQ 
* sycl_backend_compile_options_from_image_EQ 
* sycl_backend_link_options_from_image_EQ 

This patch fixes the support of the following clang command line options
in the New Offload Model. A corresponding E2E test is added. The option
list:
* -ftarget-compile-fast
* -ftarget-export-symbols
* -ftarget-register-alloc-mode