[SYCL] Invalidate handler impl after finalize
This commit changes the handler::finalize function to invalidate the
handle_impl object. This is to help ensure that further accesses to the
impl is invalid. This helps ensure that picking apart the handler and
stealing its members is internalized to finalize, allowing further
moves without concern that future code might use the members after move,
making the ownership clearer.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>