[SYCL] Fix complex tanh
The current implementation of the tanh function for the SYCL complex
extension did not always return values precise enough, neither on
device nor on host. This was caused by accumulated error from a call
to cos. This commit changes the implementation of tanh to use a
different way of calculating the result and tests it against std::tanh
for complex numbers.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>