[SYCL] Fix handling of comma separated arch value when calling llvm-offload-binary (#20130)
Because `llvm-offload-binary` uses commas to separate the key-values
pairs when specifying an image, we cannot specify a value that contains
a comma without some special handling. However, if you specify the same
key multiple times when calling `llvm-offload-binary`, the keys will be
concatenated together with commas in between. So if we want, for
example, the image to have an `arch` of `pvc,bdw`, we can call
`llvm-offload-binary` with `arch=pvc,arch=bdw` to achieve this.