[SYCL][E2E] Move AOT tests off Gen12 (#20643)
On Windows, the GPU driver split into two, one that only supports GPUs
after Gen12, and only that only supports GPUs before Gen12.

So for AOT we can't compile for BMG and Gen12 using the same `ocloc`
executable.

Instead of having to add LIT infrastructure for two `ocloc` executables
and installing two versions on all the runners, just use DG2 instead of
Gen12 for AOT tests which is supported on the new driver.

Please make sure this change doesn't invalidate the test, for example if
there is some capability difference between Gen12 and DG2 being tested
that I broke.

---------

Signed-off-by: Nick Sarnie <nick.sarnie@intel.com>