[SYCL][clang] Add attr/props warning case for multiple attributes (#20652)
This commit adds new cases to the existing tests for checking the
warning issued when there may be potential conflicts between SYCL
attributes and compile-time properties. These tests check that multiple
warning diagnostics are issued if there are multiple attributes on the
same kernel, each specifying the potentially conflicting attribute.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>