[SYCL] optimize enqueueImpKernel by inlining (#20681)
_[functional change]:_ **GlobalHandler instance move** from a static
variable declared inside a method to global scope. This **changes when
the first object is created**. Previously it would only be created if
getInstancePtr was called, now it happens at the time of loading the
library. Users pay for the instance even if they never use the global
handle.
Motivation is performance - accessing global variable declared inside a
function can be slower than to variable in global scope. Explained e.g.
[here](https://stackoverflow.com/questions/52198322/is-access-to-a-static-function-variable-slower-than-access-to-a-global-variable/52202005#52202005)

_[non-functional-change]_: **various functions** exising on hot
submit-kernel path **get inlined**

This PR is part of https://github.com/intel/llvm/pull/20668 which was
split into two. See plots in that PR desciption for performance
comparisons.