[Driver] Fix preview breaking changes option passing to offload-wrapper
When passing options to the clang-offload-wrapper, the
-fpreview-breaking-changes option needs to be passed accordingly to the
tool when -fpreview-breaking-changes is used on the command line.  This
was happening for typical SYCL enabled compilations, but when using
-fsycl-link, the usage of clang-offload-wrapper modifies the offloading
kind encountered, causing the option to not be passed.

Fixup the ability to pass -fpreview-breaking-changes for all needed uses
as well as factoring in the offload-compress options which are in the
same bucket of behaviors.