[SYCLLowerIR] Fix a hierarchical parallelism bug (#20550)
There is a bug in my earlier PR #20484. When encountering an indirect
call to a `parallel_for_work_item` function, the range of instructions
that will be subjected to the lowering pass must be closed, whereas the
current code simply skips over this instruction as its iterating the
instructions of the basic block which effectively keeps extending the
range. There is an E2E test that reflects this bug where it used to hang
before this fix but so far I'm not aware of an LLVM-IR snippet of code
that triggers it so for the moment, I'm just adding a SYCL-level test.