[NewOffloadModel]  Update clang-linker-wrapper test with issue fixes (#20697)
This pull request resolves a typo introduced in a recent modification of
`clang/test/Driver/clang-linker-wrapper.cpp`. In addition, the test
logic is improved by replacing the use of `CHK-NO-CMDS-AOT-GEN-NOT` with
`CHK-NO-CMDS-AOT-GEN` to verify the behavior where the device name
should not be passed to `sycl-post-link` when `clang-linker-wrapper` is
called with the `--gpu-tool-arg=-device arch` argument.

The previous PR introduced the mistake is
https://github.com/intel/llvm/pull/20470