[Driver][NFC] Fix test to not check for libspirv library (#20709)
Under certain testing conditions, the nvptx and amdgcn based libspirv
libraries are not generated, causing a failure in the driver test
sycl-offload-new-driver.c. Update the test to not depend on these
libraries as they are not important for the test.

Test fails with: clang: error: cannot find
'remangled-l64-signed_char.libspirv-nvptx64-nvidia-cuda.bc'; provide
path to libspirv library via '-fsycl-libspirv-path', or pass
'-fno-sycl-libspirv' to build without linking with libspirv

Test updated to use `-fno-sycl-libspirv` to prevent the inclusion of the
libs.