[SYCLLowerIR] Fix metadata format for Intel cache controls (#20753)
This patch aligns metadata with the format defined in the commit message
for

https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/c8bfc33ef8f9f273546a49d87a6612de0791e3d4.
Unfortunately, the format is still not documented in

https://github.com/KhronosGroup/SPIRV-LLVM-Translator/blob/main/docs/SPIRVRepresentationInLLVM.rst.

This patch fixes the crash in the translator tool, which tries to handle
wrong metadata format. I manually checked that the translator
successfully converts properties_cache_control.cpp test to SPIR-V format
with CacheControl decorations.