[SYCL] Fix memory access inside deallocated region when using default accessor (#20632)
This PR addresses the root cause behind #20423. When an accessor that
has not been constructed with a `handler` instance is passed to the
`set_args` function of a `handler` such as a default accessor, the
lifetime of its dynamically allocated resources can potentially be
shorter than the lifetime of the `handler` itself which causes problems,
for instance, in the `finalize` function of the `handler` where we may
end up accessing freed memory.