[SYCL] optimize enqueueImpKernel by making trace faster (#20682)
Part of https://github.com/intel/llvm/pull/20668 related to making
tracing kernels shorter.

Core optimization is based on making checking that traces are disabled
as fast & short as possible by
- use __builtin_expect indicating common path
- putting all code of non-common path in a separate not inlined function
(in cpp)
- remove "bool ResetCache" parameter and not handling its logic on
common path

So inlined check is now just "if" and one function call branch which is
a small instruction set and moreover should be optimized out by compiler
to non-common path of instructions predicted.
```
    if (__builtin_expect(SYCLConfigTrace::isTraceInMemCache(), false))
      traceKernelImpl(Msg, KernelName, isFastKernelCache);
```

By chance I did some refactor

SYCLConfig<SYCL_CACHE_TRACE> is a specialized class but code does not
utilize fact that it is a template, that is SyclConfig is never called
with generic type which evaluates to SYCL_CACHE_TRACE. Therefore this
specialization should be a simple class. This eases coding stuff in it.