[SYCL RTC] Exclude libdevice's `.o/.spv` files from resources
We only use `.bc` files from libdevice. Another potential implementation
for this was to update libdevice's CMakeLists.txt to have separate
`libdevice-bc` and `libdevice-non-bc` install components and only
install the former into `rtc-resoruce-install` directory. However, that
would be intrusive to libdevice with zero benefit there, so I decided to
do the filter on the sycl-jit's side inside `generate.py`.