[SYCL] Address Issue about Unable to Specify Kernel Properties when Reductions are Present (#20491)
Basically to fix this
[issue](https://github.com/intel/llvm/issues/16320). Note: seems that
other than certain variants (overloads) of `nd_launch` which is
explicitly mentioned in the issue, some variants of `parallel_for` with
reductions are also having the same problem. Furthermore, these
problematic `nd_launch` variants actually calls the problematic
`paralle_for` variants, so in the end only `parallel_for` variants are
modified.

Also there're some test revisions in this PR. These changes are
previously blocked by the aforementioned issue. And as of the revision
itself, the key idea here is to get rid of all
`parallel_for`/`single_task` overloads that take a property list as a
parameter (since these overloads are to be deprecated).

---------

Signed-off-by: Hu, Peisen <peisen.hu@intel.com>