[Driver][SYCL] Compile source file for integration header generation
When compiler for preprocessing only, the preprocessing step that occurs
does not generate the integration header/footer information that is used
during the host compilation step.  To generate the integration information
in this case, the driver inserts an additional compilation step that
creates the files to be consumed at the host step.

The compilation step to produce the integration information is using the
generated preprocessed file. For cases where there are build warnings that
are emitted during this step, use of -Werror will cause the compilation to
fail.  These diagnostics are typically suppressed when not performing
preprocess only compilations, but due to the fact that the build step
against the preprocessed file does not know to suppress the diagnostics
from system headers, compilation behaviors are at a disconnect.

To fix this, the behavior when creating a preprocessed file (i.e. when
we are using -E and stopping compilation), the driver will now compile the
original source file instead of the preprocessed file.